+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DEllUTY4.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-D0gEDXLV.js";import{_ as a}from"./mermaid.core-C2sazGaD.js";import"./chunk-5VM5RSS4-0_3LNjkc.js";import"./chunk-XXDRQBXY-DGGD9j-7.js";import"./chunk-POPQ4Y6H-BGDluVpO.js";import"./chunk-F27PBJKO-CiCMgOb7.js";import"./channel-Dk5SM1ID.js";import"./index-BFCDYI2c.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};